requirement.txt
fastapi
fastapi-controller
fastapi-router-controller
pymysql
uvicorn
openai
boto3
unstructured 
chromadb 
Cython 
tiktoken 
pypdf
unstructured
tabulate
pdf2image
pyreadline3
chromadb
kor
scrapy
googlesearch-python
python-dotenv